feat: Complete endpoint registration (#72)
# Recap

Endpoint registration encompasses two steps:

1. The app should request a _Client Registration Authorization_ (CRA) from the gateway via a bound service. The CRA is what we originally meant to implement as a nonce, but I decided not to call it nonce anymore because I realised it was a lot more than just a nonce.
1. The app should sign the CRA and send it back to the gateway via the PDC server. This request is called _Client Registration Request_ (CRR) and, in addition to the CRA, it also includes the endpoint's public key. When the request is valid, the PDC server issues a new certificate for the endpoint and returns it along with the gateway's certificate -- That response is known as _Client Registration_.

The CRA is valid for 5 seconds. I think should be plenty of time for the app to sign it, wake up the PDC server and `POST` the CRR, but I'd like to hear your thoughts on it.

The CRA is only meant to serve two purposes: Prevent replay attacks and mitigate the risk of an app impersonating another app (if the original CRA grantee leaks the CRA).

CRAs are bound to a specific application id. Theoretically, that means that the same app could reuse it to register multiple endpoints whilst the CRA is still valid. I think that's OK -- I don't see any security threat there.

# Summary of changes

- Rename "nonce" to CRA.
- Implement the generation of CRAs and integrate it in the bound service.
- Implement the processing of CRRs and integrate it in the PDC server. Namely:
  - Refuse invalid CRRs.
  - Register the endpoint by storing its application's id in the DB.
  - Issuing a certificate for the endpoint and return it in the client registration.

Fixes #17.